pyproject.toml
[project]
name = "g-dai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "click>=8.1.8",
    "cohere>=5.15.0",
    "docling>=2.31.0",
    "dramatiq[rabbitmq,watch]>=1.17.1",
    "fastapi-cli>=0.0.7",
    "fastapi[standard]>=0.115.12",
    "pgvector>=0.4.0",
    "python-dotenv>=1.1.0",
    "rich>=13.9.4",
    "scipy>=1.15.2",
    "streamlit>=1.44.0",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 200

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]


[tool.pytest.ini_options]
filterwarnings = ["ignore::pydantic.warnings.PydanticDeprecatedSince20"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.taskipy.tasks]
runserver = { cmd = "uvicorn src.api.app:app --reload", help = "Execute FastAPI server in development mode" }
tests = { cmd = "pytest", help = "Run all unit tests" }


